+++ test/lens_SUITE.erl
@@ -43,6 +43,11 @@
 %%
 %% pure lens interface
 -export([
+   id/1,
+   const/1,
+   hbits/1,
+   tbits/1,
+   bits/1,
    hd/1,
    hd_om/1,
    tl/1, 
@@ -102,6 +107,8 @@
 %%%----------------------------------------------------------------------------   
 all() ->
    [
+      {group, basic},
+      {group, binary},
       {group, list},
       {group, tuple},
       {group, map},
@@ -114,6 +121,12 @@ all() ->
 
 groups() ->
    [
+      {basic, [parallel],
+         [id, const]},
+
+      {binary, [parallel],
+         [hbits, tbits, bits]},
+
       {list, [parallel],
          [hd, hd_om, tl, tl_om, traverse, takewith, takewith_om]},
 
@@ -183,6 +196,51 @@ law_put_get(Lens, Value, Struct) ->
 %%          to very well behaved lenses.
 law_put_put(Lens, Value1, Value2, Expect, Struct) ->
    Expect = lens:put(Lens, Value2, lens:put(Lens, Value1, Struct)).
+
+%%%----------------------------------------------------------------------------   
+%%%
+%%% basic lenses 
+%%%
+%%%----------------------------------------------------------------------------   
+
+id(_Config) ->
+   Lens = lens:id(),
+   Value = 1,
+   law_get_put(Lens, Value),
+   law_put_get(Lens, a, Value),
+   law_put_put(Lens, a, b, b, Value).
+
+const(_Config) ->
+   Lens = lens:const(a),
+   a = lens:get(Lens, 1),
+   a = lens:put(Lens, b, 1).
+
+%%%----------------------------------------------------------------------------   
+%%%
+%%% binary lenses 
+%%%
+%%%----------------------------------------------------------------------------   
+
+hbits(_Config) ->
+   Lens = lens:hbits(8),
+   List = <<"abc">>,
+   law_get_put(Lens, List),
+   law_put_get(Lens, <<"x">>, List),
+   law_put_put(Lens, <<"x">>, <<"y">>, <<"ybc">>, List).
+
+tbits(_Config) ->
+   Lens = lens:tbits(16),
+   List = <<"abc">>,
+   law_get_put(Lens, List),
+   law_put_get(Lens, <<"x">>, List),
+   law_put_put(Lens, <<"x">>, <<"y">>, <<"aby">>, List).
+
+bits(_Config) ->
+   Lens = lens:bits(8, 8),
+   List = <<"abc">>,
+   law_get_put(Lens, List),
+   law_put_get(Lens, <<"x">>, List),
+   law_put_put(Lens, <<"x">>, <<"y">>, <<"ayc">>, List).
 
 
 %%%----------------------------------------------------------------------------